# Conflicts:
#	Domibus-MSH-angular/pom.xml
#	Domibus-MSH-api/pom.xml
#	Domibus-MSH-db/pom.xml
#	Domibus-MSH-distribution/pom.xml
#	Domibus-MSH-ext-services-delegate/pom.xml
#	Domibus-MSH-jms-activemq/pom.xml
#	Domibus-MSH-jms-spi-helper/pom.xml
#	Domibus-MSH-jms-spi/pom.xml
#	Domibus-MSH-jms-weblogic/pom.xml
#	Domibus-MSH-jms-wildfly-artemis/pom.xml
#	Domibus-MSH-logging/pom.xml
#	Domibus-MSH-soapui-tests/pom.xml
#	Domibus-MSH-swagger/pom.xml
#	Domibus-MSH-taskexecutor-tomcat/pom.xml
#	Domibus-MSH-taskexecutor-weblogic/pom.xml
#	Domibus-MSH-taskexecutor-wildfly/pom.xml
#	Domibus-MSH-test/pom.xml
#	Domibus-MSH-tomcat/pom.xml
#	Domibus-MSH-weblogic-common/pom.xml
#	Domibus-MSH-weblogic-ecas/pom.xml
#	Domibus-MSH-weblogic/pom.xml
#	Domibus-MSH-wildfly/pom.xml
#	Domibus-MSH/pom.xml
#	Domibus-authentication-dss-extension/pom.xml
#	Domibus-default-fs-plugin/pom.xml
#	Domibus-default-jms-plugin/pom.xml
#	Domibus-default-ws-plugin/pom.xml
#	Domibus-ext-model/pom.xml
#	Domibus-iam-spi/pom.xml
#	Domibus-plugin-api/pom.xml
#	Domibus-ws-stubs/pom.xml
#	pom.xml

+++ Domibus-MSH-api/src/main/java/eu/domibus/api/cluster/Command.java
@@ -23,7 +23,7 @@ public class Command {
     public static final String MESSAGE_FILTER_UPDATE = "MESSAGE_FILTER_UPDATE";
     public static final String USER_SESSION_INVALIDATION = "USER_SESSION_INVALIDATION";
 
-    private int entityId;
+    private long entityId;
     protected String commandName;
     protected String serverName;
     protected Date creationTime;
@@ -54,11 +54,11 @@ public class Command {
         this.creationTime = creationTime;
     }
 
-    public int getEntityId() {
+    public long getEntityId() {
         return entityId;
     }
 
-    public void setEntityId(int entityId) {
+    public void setEntityId(long entityId) {
         this.entityId = entityId;
     }
