+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/multitenancy/MultiDomainPModeProvider.java
@@ -274,7 +274,7 @@ public class MultiDomainPModeProvider extends PModeProvider {
     }
 
     @Override
-    public byte[] getPModeFile(int id) {
+    public byte[] getPModeFile(long id) {
         return getCurrentPModeProvider().getPModeFile(id);
     }
 
@@ -289,7 +289,7 @@ public class MultiDomainPModeProvider extends PModeProvider {
     }
 
     @Override
-    public void removePMode(int id) {
+    public void removePMode(long id) {
         getCurrentPModeProvider().removePMode(id);
     }
